IGNITE-17149 after review #1.0
