+++ modules/page-memory/src/integrationTest/java/org/apache/ignite/internal/pagememory/tree/ItBplusTreePageMemoryTest.java
@@ -2731,6 +2731,7 @@ public abstract class ItBplusTreePageMemoryTest extends BaseIgniteAbstractTest {
      * Returns number of acquired pages.
      */
     protected abstract long acquiredPages();
+
     /**
      * Long leaf.
      */
